refactor: nft list
